🔨 remove one use global functions => local fn
